requirements.txt
# Agency Management System - Python Dependencies
# ================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Agent & AI Orchestration
openai==1.12.0
anthropic==0.18.0
groq==0.4.0
httpx==0.26.0

# Async & Concurrency
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Workflow Engine
celery==5.3.6

# Terminal UI
rich==13.7.0
typer==0.9.0

# WebSocket & Real-time
websockets==12.0
python-socketio==5.11.0

# Monitoring & Metrics
psutil==5.9.8
structlog==24.1.0

# Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pyyaml==6.0.1
jinja2==3.1.3
tenacity==8.2.3
croniter==2.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
